refactor(rubrics): Group rubrics requests
