--- src/rubrics/infrastructure/requests/add_objective_to_rubric_request.go
@@ -1,5 +1,0 @@
-package requests
-
-type AddObjectiveToRubricRequest struct {
-	Description string `json:"description" validate:"required,min=8,max=510"`
-}
