--- src/rubrics/infrastructure/requests/create_rubric_request.go
@@ -1,5 +1,0 @@
-package requests
-
-type CreateRubricRequest struct {
-	Name string `json:"name" validate:"required,min=4,max=96"`
-}
