+++ src/rubrics/infrastructure/requests/objectives_requests.go
@@ -1,5 +1,9 @@
 package requests
 
+type AddObjectiveToRubricRequest struct {
+	Description string `json:"description" validate:"required,min=8,max=510"`
+}
+
 type UpdateObjectiveRequest struct {
 	Description string `json:"description" validate:"required,min=8,max=510"`
 }
