+++ src/rubrics/infrastructure/requests/rubrics_requests.go
@@ -1,5 +1,9 @@
 package requests
 
+type CreateRubricRequest struct {
+	Name string `json:"name" validate:"required,min=4,max=96"`
+}
+
 type UpdateRubricNameRequest struct {
 	Name string `json:"name" validate:"required,min=4,max=96"`
 }
